Added vision to chat. Minor refactoring. Added uv.lock and fin_data
