+++ .gitignore
@@ -10,4 +10,3 @@ wheels/
 .venv
 .env
 serviceAccountKey.json
-uv.lock
