+++ models.py
@@ -1,3 +1,5 @@
+from typing import Optional
+
 from pydantic import BaseModel
 
 
@@ -5,13 +7,19 @@ class ChatRequest(BaseModel):
     # userId: str
     message: str
     chatSessionId: str
-    
+
+class ChatRequestImage(BaseModel):
+    # userId: str
+    message: str
+    chatSessionId: Optional[str] = None
+    imageUrl: Optional[str] = None
+
 class StartRequest(BaseModel):
-    userId: str
+    # userId: str
     level: str
 
 class AnswerRequest(BaseModel):
-    userId: str
+    # userId: str
     sessionId: str
     answer: str
