+++ requirements.txt
@@ -1,0 +1,257 @@
+# This file was autogenerated by uv via the following command:
+#    uv pip compile pyproject.toml -o requirements.txt
+annotated-types==0.7.0
+    # via pydantic
+anyio==4.9.0
+    # via
+    #   google-genai
+    #   httpx
+    #   starlette
+    #   watchfiles
+beautifulsoup4==4.13.3
+    # via google
+cachecontrol==0.14.2
+    # via firebase-admin
+cachetools==5.5.2
+    # via google-auth
+certifi==2025.1.31
+    # via
+    #   httpcore
+    #   httpx
+    #   requests
+cffi==1.17.1
+    # via cryptography
+charset-normalizer==3.4.1
+    # via requests
+click==8.1.8
+    # via
+    #   rich-toolkit
+    #   typer
+    #   uvicorn
+cryptography==44.0.2
+    # via pyjwt
+dnspython==2.7.0
+    # via email-validator
+email-validator==2.2.0
+    # via fastapi
+fastapi==0.115.12
+    # via tutor-api-gdg (pyproject.toml)
+fastapi-cli==0.0.7
+    # via fastapi
+firebase-admin==6.7.0
+    # via tutor-api-gdg (pyproject.toml)
+google==3.0.0
+    # via tutor-api-gdg (pyproject.toml)
+google-ai-generativelanguage==0.6.15
+    # via google-generativeai
+google-api-core==2.24.2
+    # via
+    #   firebase-admin
+    #   google-ai-generativelanguage
+    #   google-api-python-client
+    #   google-cloud-core
+    #   google-cloud-firestore
+    #   google-cloud-storage
+    #   google-generativeai
+google-api-python-client==2.165.0
+    # via
+    #   firebase-admin
+    #   google-generativeai
+google-auth==2.38.0
+    # via
+    #   google-ai-generativelanguage
+    #   google-api-core
+    #   google-api-python-client
+    #   google-auth-httplib2
+    #   google-cloud-core
+    #   google-cloud-firestore
+    #   google-cloud-storage
+    #   google-genai
+    #   google-generativeai
+google-auth-httplib2==0.2.0
+    # via google-api-python-client
+google-cloud-core==2.4.3
+    # via
+    #   google-cloud-firestore
+    #   google-cloud-storage
+google-cloud-firestore==2.20.1
+    # via firebase-admin
+google-cloud-storage==3.1.0
+    # via firebase-admin
+google-crc32c==1.7.0
+    # via
+    #   google-cloud-storage
+    #   google-resumable-media
+google-genai==1.7.0
+    # via tutor-api-gdg (pyproject.toml)
+google-generativeai==0.8.4
+    # via tutor-api-gdg (pyproject.toml)
+google-resumable-media==2.7.2
+    # via google-cloud-storage
+googleapis-common-protos==1.69.2
+    # via
+    #   google-api-core
+    #   grpcio-status
+grpcio==1.71.0
+    # via
+    #   google-api-core
+    #   grpcio-status
+grpcio-status==1.71.0
+    # via google-api-core
+h11==0.14.0
+    # via
+    #   httpcore
+    #   uvicorn
+httpcore==1.0.7
+    # via httpx
+httplib2==0.22.0
+    # via
+    #   google-api-python-client
+    #   google-auth-httplib2
+httptools==0.6.4
+    # via uvicorn
+httpx==0.28.1
+    # via
+    #   fastapi
+    #   google-genai
+idna==3.10
+    # via
+    #   anyio
+    #   email-validator
+    #   httpx
+    #   requests
+itsdangerous==2.2.0
+    # via fastapi
+jinja2==3.1.6
+    # via fastapi
+markdown-it-py==3.0.0
+    # via rich
+markupsafe==3.0.2
+    # via jinja2
+mdurl==0.1.2
+    # via markdown-it-py
+msgpack==1.1.0
+    # via cachecontrol
+numpy==2.2.4
+    # via pandas
+orjson==3.10.16
+    # via fastapi
+pandas==2.2.3
+    # via tutor-api-gdg (pyproject.toml)
+proto-plus==1.26.1
+    # via
+    #   google-ai-generativelanguage
+    #   google-api-core
+    #   google-cloud-firestore
+protobuf==5.29.4
+    # via
+    #   google-ai-generativelanguage
+    #   google-api-core
+    #   google-cloud-firestore
+    #   google-generativeai
+    #   googleapis-common-protos
+    #   grpcio-status
+    #   proto-plus
+pyasn1==0.6.1
+    # via
+    #   pyasn1-modules
+    #   rsa
+pyasn1-modules==0.4.1
+    # via google-auth
+pycparser==2.22
+    # via cffi
+pydantic==2.10.6
+    # via
+    #   fastapi
+    #   google-genai
+    #   google-generativeai
+    #   pydantic-extra-types
+    #   pydantic-settings
+pydantic-core==2.27.2
+    # via pydantic
+pydantic-extra-types==2.10.3
+    # via fastapi
+pydantic-settings==2.8.1
+    # via fastapi
+pygments==2.19.1
+    # via rich
+pyjwt==2.10.1
+    # via firebase-admin
+pyparsing==3.2.2
+    # via httplib2
+python-dateutil==2.9.0.post0
+    # via pandas
+python-dotenv==1.0.1
+    # via
+    #   tutor-api-gdg (pyproject.toml)
+    #   pydantic-settings
+    #   uvicorn
+python-multipart==0.0.20
+    # via fastapi
+pytz==2025.1
+    # via pandas
+pyyaml==6.0.2
+    # via
+    #   fastapi
+    #   uvicorn
+requests==2.32.3
+    # via
+    #   tutor-api-gdg (pyproject.toml)
+    #   cachecontrol
+    #   google-api-core
+    #   google-cloud-storage
+    #   google-genai
+rich==13.9.4
+    # via
+    #   rich-toolkit
+    #   typer
+rich-toolkit==0.13.2
+    # via fastapi-cli
+rsa==4.9
+    # via google-auth
+shellingham==1.5.4
+    # via typer
+six==1.17.0
+    # via python-dateutil
+sniffio==1.3.1
+    # via anyio
+soupsieve==2.6
+    # via beautifulsoup4
+starlette==0.46.1
+    # via fastapi
+tqdm==4.67.1
+    # via google-generativeai
+typer==0.15.2
+    # via fastapi-cli
+typing-extensions==4.12.2
+    # via
+    #   beautifulsoup4
+    #   fastapi
+    #   google-genai
+    #   google-generativeai
+    #   pydantic
+    #   pydantic-core
+    #   pydantic-extra-types
+    #   rich-toolkit
+    #   typer
+tzdata==2025.2
+    # via pandas
+ujson==5.10.0
+    # via fastapi
+uritemplate==4.1.1
+    # via google-api-python-client
+urllib3==2.3.0
+    # via requests
+uvicorn==0.34.0
+    # via
+    #   tutor-api-gdg (pyproject.toml)
+    #   fastapi
+    #   fastapi-cli
+uvloop==0.21.0
+    # via uvicorn
+watchfiles==1.0.4
+    # via uvicorn
+websockets==15.0.1
+    # via
+    #   google-genai
+    #   uvicorn
